즐겨찾기 & 길찾기 추적 로그 추가
