+++ OilPrice-Where/Scenes/Favorite/View/FavoriteCollectionViewCell.swift
@@ -230,16 +230,6 @@ class FavoriteCollectionViewCell: UICollectionViewCell {
             .tapGesture()
             .when(.recognized)
             .subscribe(with: self, onNext: { owner, _ in
-                let event = "tap_favorite_copy_address"
-                let parameters = [
-                    "file": #file,
-                    "function": #function,
-                    "eventDate": DefaultData.shared.currentTime
-                ]
-                
-                Analytics.setUserProperty("ko", forName: "country")
-                Analytics.logEvent(event, parameters: parameters)
-                
                 guard let valueString = owner.addressHStackView.valueLabel.text else { return }
                 UIPasteboard.general.string = valueString
 
@@ -258,15 +248,6 @@ class FavoriteCollectionViewCell: UICollectionViewCell {
             .tapGesture()
             .when(.recognized)
             .subscribe(with: self, onNext: { owner, _ in
-                let event = "tap_favorite_copy_phone"
-                let parameters = [
-                    "file": #file,
-                    "function": #function,
-                    "eventDate": DefaultData.shared.currentTime
-                ]
-                
-                Analytics.setUserProperty("ko", forName: "country")
-                Analytics.logEvent(event, parameters: parameters)
                 guard let valueString = owner.phoneNumberHStackView.valueLabel.text,
                       let url = URL(string: "tel:" + valueString),
                       UIApplication.shared.canOpenURL(url) else { return }
