+++ OilPrice-Where/Scenes/Main/List/MainListVC.swift
@@ -13,6 +13,7 @@ import RxSwift
 import RxCocoa
 import NMapsMap
 import SCLAlertView
+import FirebaseAnalytics
 
 protocol MainListVCDelegate: AnyObject {
     func touchedCell(info: GasStation)
@@ -206,6 +207,16 @@ extension MainListVC: UICollectionViewDelegate, UICollectionViewDelegateFlowLayo
 extension MainListVC: GasStationCellDelegate {
     // 즐겨찾기 설정 및 해제
     func touchedFavoriteButton(id: String?) {
+        let event = "tap_list_favorite"
+        let parameters = [
+            "file": #file,
+            "function": #function,
+            "eventDate": DefaultData.shared.currentTime
+        ]
+        
+        Analytics.setUserProperty("ko", forName: "country")
+        Analytics.logEvent(event, parameters: parameters)
+        
         let faovorites = DefaultData.shared.favoriteSubject.value
         guard let _id = id, faovorites.count < 6 else { return }
         let isDeleted = faovorites.contains(_id)
@@ -222,6 +233,16 @@ extension MainListVC: GasStationCellDelegate {
     }
     
     func touchedDirectionButton(info: GasStation?) {
+        let event = "tap_list_navigation"
+        let parameters = [
+            "file": #file,
+            "function": #function,
+            "eventDate": DefaultData.shared.currentTime
+        ]
+        
+        Analytics.setUserProperty("ko", forName: "country")
+        Analytics.logEvent(event, parameters: parameters)
+        
         requestDirection(station: info)
     }
 }
